+++ SqlDataScripter/Options.cs
@@ -12,19 +12,19 @@ namespace SqlDataScripter
         [Option('s', "server", HelpText = "SQL Server instance.", Default = ".")]
         public string Server { get; set; }
 
-        [Option('u', "user", Required = false, HelpText = "SQL user name.")]
-        public string UserName { get; set; }
-
-        [Option('p', "password", Required = false, HelpText = "SQL user password.")]
-        public string Password { get; set; }
-
         [Option('d', "database", Required = true, HelpText = "Script database name.")]
         public string Database { get; set; }
 
-        [Option('o', "output", Required = false, HelpText = "Output file name.")]
+        [Option('u', "user", HelpText = "SQL user name.")]
+        public string UserName { get; set; }
+
+        [Option('p', "password", HelpText = "SQL user password.")]
+        public string Password { get; set; }
+
+        [Option('o', "output", HelpText = "Output file name.")]
         public string OutputFileName { get; set; }
 
-        [Option('t', "table", Required = false)]
+        [Option('t', "table", HelpText = "Table name(s) to script. Script all tables in database if empty.")]
         public IEnumerable<string> Tables { get; set; }
 
         [Option("schema", HelpText = "Script table schema.")]
@@ -33,13 +33,25 @@ namespace SqlDataScripter
         [Option("data", HelpText = "Script table data.")]
         public bool ScriptData { get; set; }
 
-        [Option]
+        [Option(HelpText = "Output file encoding.")]
+        public string Encoding { get; set; }
+
+        [Option(HelpText = "Append to file not overwrite.")]
+        public bool AppendToFile { get; set; }
+
+        [Option(HelpText = "Disables GO terminator in script.")]
+        public bool NoBatchTerminator { get; set; }
+
+        [Option(HelpText = "Add an existence check to script.")]
         public bool IncludeIfNotExists { get; set; }
 
-        [Option(Hidden = false)]
+        [Option(HelpText = "Include a header containing information about the object being scripted.")]
         public bool IncludeHeaders { get; set; }
 
-        [Option(HelpText = "Suppress any messages.")]
+        [Option(HelpText = "Script table indexes.")]
+        public bool Indexes { get; set; }
+
+        [Option('q', "quiet", HelpText = "Suppress any informational messages.")]
         public bool Quiet { get; set; }
 
     }